source/backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
psycopg2-binary==2.9.10
boto3==1.37.28
pydantic==2.11.3
